V2
Self-updates now working through the SuperBLT.
Added players-in-lobby amount indicator.
Added info about lobby presets, such as: Reputation limit, Kick option, Prompt and etc.
Added option for resize the windows of the host info.
Fixed an issue when you could't open a server job with controller. (yet it will be an old ui till you using a gamepad, ill try to do something in the future with it)